requirements.txt
fastapi==0.68.1
pydantic==1.8.2
uvicorn==0.15.0
psycopg2-binary==2.9.3
hubspot-api-client
clickup
simplejson
requests==2.26.0
sqlalchemy==1.4.25
python-dotenv
python-decouple